alias-detection-code
As a group of students, we recently uncovered a notable vulnerability in a software system pertaining to the usage of aliases and aliases of aliases, which could potentially bypass security measures intended to restrict function calls. Upon identifying this flaw, we undertook the task of developing a correction to address this issue. Our approach involved enhancing the existing code validation framework to accurately track and resolve aliases and their derivatives, ensuring that all function calls, variable assignments, and module imports are checked against a predefined list of allowed or blocked entities. This correction not only mitigates the risk associated with the vulnerability but also reinforces the system's overall security against similar exploits in the future.

this was done by : 
Martinet theo 
Risch guillaume
El hammouti nissrine
El haitamy raihane
Sedeno evan 
Rousset theo